requirements.txt
fastapi
uvicorn
python-multipart
requests
openai
fpdf2
